move color resolver to another module
